backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.17
PyPDF2==3.0.1
python-docx==1.1.2
email-validator==2.2.0
